requirements.txt
fastapi==0.95.2
uvicorn==0.23.1
python-dotenv==1.0.0
groq==0.13.0
python-multipart==0.0.6
